+++ src/AddressBook/ABEnums.cs
@@ -37,14 +37,14 @@ namespace AddressBook {
 	[Deprecated (PlatformName.iOS, 9, 0, message : "Use the 'Contacts' API instead.")]
 	[Native]
 	[ErrorDomain ("ABAddressBookErrorDomain")]
-	public enum ABAddressBookError : nint {
+	public enum ABAddressBookError : long {
 		OperationNotPermittedByStore = 0,
 		OperationNotPermittedByUserError
 	}
 
 	[Deprecated (PlatformName.iOS, 9, 0, message : "Use the 'Contacts' API instead.")]
 	[Native]
-	public enum ABAuthorizationStatus : nint {
+	public enum ABAuthorizationStatus : long {
 		NotDetermined = 0,
 		Restricted,
 		Denied,
