+++ src/CoreAnimation/CAEnums.cs
@@ -51,7 +51,7 @@ namespace CoreAnimation {
 	[NoWatch] // headers not updated
 	[iOS (11,0)][TV (11,0)][Mac (10,13)]
 	[Native][Flags]
-	public enum CACornerMask : nuint {
+	public enum CACornerMask : ulong {
 		MinXMinYCorner = 1 << 0,
 		MaxXMinYCorner = 1 << 1,
 		MinXMaxYCorner = 1 << 2,
