+++ src/NewsstandKit/Enums.cs
@@ -14,7 +14,7 @@ namespace NewsstandKit {
 	// untyped enum -> NKIssue.h
 	// note: iOS 8 beta (2) has it defined as an NSInteger, filed as radar 17564957 with Apple
 	[Native]
-	public enum NKIssueContentStatus : nint {
+	public enum NKIssueContentStatus : long {
 		None, Downloading, Available
 	}
 }
