+++ src/NotificationCenter/NCEnums.cs
@@ -12,7 +12,7 @@ namespace NotificationCenter {
 
 	[iOS (8,0)][Mac (10,10)]
 	[Native]
-	public enum NCUpdateResult : nuint {
+	public enum NCUpdateResult : ulong {
 		NewData,
 		NoData,
 		Failed
@@ -20,7 +20,7 @@ namespace NotificationCenter {
 
 	[iOS (10,0)][NoMac]
 	[Native]
-	public enum NCWidgetDisplayMode : nint {
+	public enum NCWidgetDisplayMode : long {
 		Compact,
 		Expanded
 	}
